New change
Added white dots to the background in order to make it more space themed. Made the invaders more hostile i.e. make them faster as all of the group are shot when they regenerate they get faster. We've added sounds for the defender once it shoots lasers and added a sound once an invader is shot. There is a game over screen once the invader touches the defender. We will now add a restart button, power ups and a way to remember the high score previously.